Follow up changes 1
