--- ui/src/app/job-list/table/additional-columns.config.ts
@@ -1,12 +1,0 @@
-export const dsubColumns = [
-  {header: 'Owner', key: 'user-id'},
-  {header: 'Status Detail', key: 'status-detail'}
-];
-
-// TODO: Provide the labels that should be used with Cromwell here
-export const cromwellColumns = [];
-
-export class LabelColumn {
-  header: string;
-  key: string;
-}
