+++ ui/src/environments/additional-columns.config.ts
@@ -1,0 +1,4 @@
+export const dsubAdditionalColumns = [
+  {header: 'Owner', key: 'user-id'},
+  {header: 'Status Detail', key: 'status-detail'}
+];
