+++ ui/src/environments/environment.ts
@@ -9,5 +9,5 @@ export const environment = {
   production: false,
   requiresAuth: false,
   scope: '',
-  backendProvider: null,
+  additionalColumns: null,
 };
